+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Hi! This is where we document all notable changes, including bug fixes, enhancements, and dependency updates.
 Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](http://semver.org/) format.
 
+## 0.0.10 2024-11-17
+
+### Maintenance
+
+- Updated Svelte and Supabase point releases.
+
 ## 0.0.9 2024-11-10
 
 ### Added